pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "anamnesis"
version = "0.1.0"
description = "4D Strategic Memory Engine for Autonomous AI Agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Dr. Shweta Patel"},
]
keywords = ["ai", "memory", "agents", "rag", "strategic-memory", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.6",
    "numpy>=1.26.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
voyage = ["voyageai>=0.3.0"]
local = ["sentence-transformers>=3.0"]
anthropic = ["anthropic>=0.40.0"]
openai = ["openai>=1.0"]
mcp = ["mcp>=1.0"]
all = [
    "voyageai>=0.3.0",
    "anthropic>=0.40.0",
    "sentence-transformers>=3.0",
    "mcp>=1.0",
]

[project.urls]
Homepage = "https://github.com/shwetapateldr/anamnesis"
Repository = "https://github.com/shwetapateldr/anamnesis"
Issues = "https://github.com/shwetapateldr/anamnesis/issues"

[tool.setuptools.packages.find]
include = ["anamnesis*"]
